Flatbuffer protocols and make_fb.py,edit .gitignore

+++ src/shared/common.h
@@ -79,7 +79,7 @@
     #define NWAPIEXPORT __attribute__((visibility("default")))
 #endif
 
-constexpr unsigned int NEWorldVersion = 41u;
+constexpr unsigned short NEWorldVersion = 41u;
 constexpr const char* CopyrightString = R"(
 NEWorld  Copyright (C) 2016  NEWorld Team
 This program comes with ABSOLUTELY NO WARRANTY.
